requirements.txt
Package             Version
------------------- -------
asgiref             3.4.1
backports.zoneinfo  0.2.1
Django              4.0
django-filter       21.1
djangorestframework 3.13.1
flake8              4.0.1
importlib-metadata  4.8.3
Markdown            3.3.6
mccabe              0.6.1
pip                 21.3.1
pkg_resources       0.0.0
psycopg2-binary     2.9.2
pycodestyle         2.8.0
pyflakes            2.4.0
pytz                2021.3
setuptools          44.0.0
sqlparse            0.4.2
zipp                3.6.0
